+++ backend/app/searchHistory/routes.py
@@ -11,7 +11,6 @@ from . import searchHistory_BP
 def addSearch():
     data = request.get_json()
     subreddit = data.get("subreddit")
-    subreddit = subreddit.lower()
     sentimentKeywords = data.get("sentimentKeywords")
     startDate = data.get("startDate")
     endDate = data.get("endDate")
