+++ frontend/src/data.jsx
@@ -1,4 +1,7 @@
 import React, { useEffect, useState, useRef } from 'react';
+import { ToastContainer } from 'react-toastify';
+import { Chart as ChartJS, CategoryScale, LinearScale, PointElement, LineElement, Title, Tooltip, Legend, BarElement } from 'chart.js';
+import { Bar } from 'react-chartjs-2';
 import ReactDatePicker from 'react-datepicker';
 import 'react-datepicker/dist/react-datepicker.css';
 import $ from 'jquery';
@@ -12,35 +15,41 @@ import * as Arrow from 'apache-arrow';
 import 'pdfmake/build/vfs_fonts.js';
 import 'datatables.net-buttons/js/buttons.html5.min';
 import 'datatables.net-buttons/js/buttons.print.min';
-import TopicTablesContainer from './TopicTablesContainer';
-import handleNotify from './toast';
-import SearchSuggestions from './SearchSuggestions';
-import { ToastContainer } from 'react-toastify';
-import { Chart as ChartJS, CategoryScale, LinearScale, PointElement, LineElement, Title, Tooltip, Legend, BarElement } from 'chart.js';
-import { Bar } from 'react-chartjs-2';
+import TopicTablesContainer from './Components/TopicTablesContainer';
+import handleNotify from './Components/Toast';
+import SearchSuggestions from './Components/SearchSuggestions';
+import SearchHistory from './Components/SearchHistory';
+
 
 ChartJS.register(CategoryScale, LinearScale, BarElement, Title, Tooltip, Legend);
 
 function Data() {
+  // Main Search Components
   const [subreddit, setSubreddit] = useState('TrigeminalNeuralgia');
   const [sentimentKeywords, setSentimentKeywords] = useState('');
   const [startDate, setStartDate] = useState(new Date(2024, 11, 2)); // December 2, 2024
   const [endDate, setEndDate] = useState(new Date(2024, 11, 31));     // December 31, 2024
   const [searchTerms, setSearchTerms] = useState([]);
   const [email] = useState(localStorage.getItem('email'));
-  const [error, setError] = useState(null);
-  const [searchError, setSearchError] = useState(null);
   const [option, setOption] = useState("reddit_submissions, reddit_comments");
-  const [clusteringResults, setClusteringResults] = useState(null);
-  const [loadingSentiment, setLoadingSentiment] = useState(false);
-  const [searchData, setSearchData] = useState([]);
-  const [dataMessage, setDataMessage] = useState(false);
-  const [subredditIcon, setSubredditIcon] = useState(null);
-  // Add these new state variables at the top along with your others.
   const [includeSubmissions, setIncludeSubmissions] = useState(true);
   const [includeComments, setIncludeComments] = useState(true);
   const [debouncedSubreddit, setDebouncedSubreddit] = useState(subreddit);
   const isOptionValid = includeSubmissions || includeComments;
+  const [error, setError] = useState(null);
+  const [dataMessage, setDataMessage] = useState(false);
+  const [subredditIcon, setSubredditIcon] = useState(null);
+  const prevSubredditRef = useRef(null);
+  // Add these new state variables at the top along with your others.
+  const [clusteringResults, setClusteringResults] = useState(null);
+  const [loadingSentiment, setLoadingSentiment] = useState(false);
+
+  // Search History Variables
+  const [isDeleting, setIsDeleting] = useState(false);
+  const [searchData, setSearchData] = useState([]);
+  const historyInitializedRef = useRef(false);
+  const [searchError, setSearchError] = useState(null);
+  const [dataLoadMessage, setDataLoadMessage] = useState(false);
 
   // This ref controls whether the main results DataTable should fetch data
   const tableInitializedRef = useRef(false);
@@ -56,142 +65,12 @@ function Data() {
   useEffect(() => { optionRef.current = option; }, [option]);
 
 
-
-  // Fetch search history immediately on mount.
-  useEffect(() => {
-    const fetchSearchHistory = async () => {
-      try {
-        const response = await fetch(`/api/get_search/${encodeURIComponent(email)}`);
-
-        if (!response.ok) {
-          setSearchError('Search history failed to fetch from Flask.');
-        }
-        else {
-          setSearchError(null);
-        }
-
-        const data = await response.json();
-        setSearchData(data.search_history || []);
-      } catch (error) {
-        setSearchError(error.message);
-      }
-    };
-    fetchSearchHistory();
-  }, [email]);
-
   useEffect(() => {
     const handler = setTimeout(() => {
       setDebouncedSubreddit(subreddit);
     }, 1000);
     return () => clearTimeout(handler);
   }, [subreddit]);
-
-  const AddSearch = async () => {
-    try {
-      const response = await fetch('/api/add_search', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({
-          subreddit: `r/${subreddit}`,
-          sentimentKeywords: sentimentKeywords,
-          startDate: startDate,
-          endDate: endDate,
-          option: option,
-          email: email,
-        }),
-      });
-
-      if (!response.ok) {
-        setError('Failed to submit and add search.');
-      }
-      else {
-        setError(null);
-      }
-
-      const data = await response.json();
-      //console.log('Search Query:', data);
-      // Refresh search history after adding a search.
-      const response2 = await fetch(`/api/get_search/${encodeURIComponent(email)}`);
-      const data2 = await response2.json();
-      setSearchData(data2.search_history || []);
-    } catch (error) {
-      setSearchError(error.message);
-    }
-  };
-
-  const [isDeleting, setIsDeleting] = useState(false);
-  const historyInitializedRef = useRef(false);
-
-  const RemoveSearch = async (searchId) => {
-    try {
-      setSearchData((prevData) => prevData.filter((item) => item.id !== searchId));
-      setIsDeleting(true);
-
-      const response = await fetch(`/api/remove_search/${searchId}`, {
-        method: 'DELETE',
-      });
-
-      if (!response.ok) {
-        setSearchError('Failed to remove search ${searchId}.');
-      }
-      else {
-        setSearchError(null);
-      }
-
-      //console.log(`Search ID ${searchId} successfully removed from backend.`);
-    } catch (error) {
-      setSearchError(error.message);
-    } finally {
-      const response = await fetch(`/api/get_search/${encodeURIComponent(email)}`);
-
-      if (!response.ok) {
-        setSearchError('Failed to fetch updated search history.');
-      }
-      else {
-        setSearchError(null);
-      }
-
-      const data = await response.json();
-      setSearchData(data.search_history || []);
-      setIsDeleting(false);
-    }
-  };
-
-  const ClearAllSearch = async () => {
-    const button = document.getElementById("clear-all-btn");
-
-    try {
-      button.disabled = true;
-      setIsDeleting(true);
-      const response = await fetch(`/api/clear_all/${encodeURIComponent(email)}`, {
-        method: 'DELETE',
-      });
-
-      if (!response.ok) {
-        setSearchError("Failed to clear all searches.");
-      }
-      else {
-        setSearchError(null);
-      }
-
-      setSearchData([]);
-    } catch (error) {
-    } finally {
-      const response = await fetch(`/api/get_search/${encodeURIComponent(email)}`);
-
-      if (!response.ok) {
-        setSearchError("Failed to fetch updated search history.");
-      }
-      else {
-        setSearchError(null);
-      }
-
-      const data = await response.json();
-      setSearchData(data.search_history || []);
-      setIsDeleting(false);
-      button.disabled = false;
-    }
-  };
 
   const fetchArrowData = async (start, length, draw, searchValue, setError) => {
     try {
@@ -287,217 +166,6 @@ function Data() {
     };
   };
 
-
-
-  // Initialize search history DataTable when searchData changes.
-  useEffect(() => {
-    //console.log(searchData);
-    if (searchData.length === 0) {
-      document.getElementById("clear-all-container").style.display = 'none';
-    } else {
-      document.getElementById("clear-all-container").style.display = 'block';
-    }
-
-    if (!searchData.length) {
-      if ($.fn.DataTable.isDataTable("#search-history-table")) {
-        $("#search-history-table").DataTable().clear().draw();
-      } else {
-        $("#search-history-table").DataTable({
-          data: [],
-          columns: [
-            { data: "subreddit", title: "Subreddit" },
-            {
-              data: "sentimentKeywords",
-              title: "Keywords",
-              render: function (data) {
-                return `<span class="truncate-text" title="${data}">${data}</span>`;
-              }
-            },
-            {
-              data: "startDate",
-              title: "Start Date",
-              render: function (data) {
-                return new Date(data).toLocaleDateString('en-US');
-              }
-            },
-            {
-              data: "endDate",
-              title: "End Date",
-              render: function (data) {
-                return new Date(data).toLocaleDateString('en-US');
-              }
-            },
-            {
-              data: "option",
-              title: "Post Option",
-              render: function (data) {
-                if (!data) return "";
-
-                const options = data.split(",").map(opt => opt.trim());
-                const labels = options.map(opt => {
-                  if (opt === "reddit_comments") return "Comments";
-                  if (opt === "reddit_submissions") return "Submissions";
-                  return opt;
-                });
-
-                return labels.join(", ");
-              }
-            },
-            { data: "created_utc", title: "Created UTC" },
-            {
-              data: null,
-              title: "Actions",
-              render: function (data, type, row) {
-                const searchButton = `<a href="#" class="btn btn-primary go-to-btn" data-search-id="${row.search_id}" data-subreddit="${row.subreddit}" data-sentiment="${row.sentimentKeywords}" data-start-date="${row.startDate}" data-end-date="${row.endDate}" data-option="${row.option}" disabled>Go To Search</a>`;
-                const deleteButton = `<button class="btn btn-danger delete-btn" data-search-id="${row.search_id}" disabled>Delete</button>`;
-                return searchButton + " " + deleteButton;
-              }
-            }
-          ],
-          order: [[5, 'desc']],
-          paging: false,
-          searching: false,
-          ordering: false,
-          info: false,
-          autoWidth: false,
-          language: {
-            emptyTable: "No search history available."
-          },
-          dom: "<'row'<'col-sm-12 col-md-6'l><'col-sm-12 col-md-6'f>>" +
-            "<'row'<'col-sm-12'tr>>" +
-            "<'row'<'col-sm-12 col-md-5'i><'col-sm-12 col-md-7'p>>"
-        });
-      }
-      return;
-    }
-
-    if ($.fn.DataTable.isDataTable("#search-history-table")) {
-      $("#search-history-table").DataTable().destroy();
-    }
-
-    $("#search-history-table").DataTable({
-      data: searchData,
-      columns: [
-        { data: "subreddit", title: "Subreddit" },
-        {
-          data: "sentimentKeywords",
-          title: "Keywords",
-          render: function (data) {
-            return `<span class="truncate-text" title="${data}">${data}</span>`;
-          }
-        },
-        {
-          data: "startDate",
-          title: "Start Date",
-          render: function (data) {
-            return new Date(data).toLocaleDateString('en-US');
-          }
-        },
-        {
-          data: "endDate",
-          title: "End Date",
-          render: function (data) {
-            return new Date(data).toLocaleDateString('en-US');
-          }
-        },
-        {
-          data: "option",
-          title: "Post Option",
-          render: function (data) {
-            if (!data) return "";
-
-            const options = data.split(",").map(opt => opt.trim());
-            const labels = options.map(opt => {
-              if (opt === "reddit_comments") return "Comments";
-              if (opt === "reddit_submissions") return "Submissions";
-              return opt;
-            });
-
-            return labels.join(", ");
-          }
-        },
-        { data: "created_utc", title: "Created UTC" },
-        {
-          data: null,
-          title: "Actions",
-          render: function (data, type, row) {
-            const searchButton = `<a href="#" class="btn btn-primary go-to-btn" data-search-id="${row.search_id}" data-subreddit="${row.subreddit}" data-sentiment="${row.sentimentKeywords}" data-start-date="${row.startDate}" data-end-date="${row.endDate}" data-option="${row.option}" disabled>Go To Search</a>`;
-            const deleteButton = `<button class="btn btn-danger delete-btn" data-search-id="${row.search_id}">Delete</button>`;
-            return searchButton + " " + deleteButton;
-          }
-        }
-      ],
-      order: [[5, 'desc']],
-      paging: true,
-      searching: false,
-      ordering: true,
-      info: true,
-      autoWidth: false,
-      language: {
-        emptyTable: "No search history available."
-      },
-      dom: "<'row'<'col-sm-12 col-md-6'l><'col-sm-12 col-md-6'f>>" +
-        "<'row'<'col-sm-12'tr>>" +
-        "<'row'<'col-sm-12 col-md-5'i><'col-sm-12 col-md-7'p>>"
-    });
-
-    if (!historyInitializedRef.current) {
-      historyInitializedRef.current = true;
-
-      $(document).on("click", ".delete-btn", function () {
-        const searchId = $(this).data("search-id");
-        RemoveSearch(searchId);
-      });
-    }
-
-    $(document).on("click", ".go-to-btn", function () {
-      const subredditGo = $(this).data("subreddit").replace(/^r\//, '');
-      const sentimentGo = $(this).data("sentiment");
-      const startDateGo = new Date($(this).data("start-date"));
-      const endDateGo = new Date($(this).data("end-date"));
-      const optionGo = $(this).data("option");
-      const searchValue = `${subredditGo} ${sentimentGo} ${startDateGo.toISOString()} ${endDateGo.toISOString()} ${optionGo}`;
-
-      if (optionGo.includes('reddit_submissions')) {
-        setIncludeSubmissions(true);
-      }
-      else {
-        setIncludeSubmissions(false);
-      }
-
-      if (optionGo.includes('reddit_comments')) {
-        setIncludeComments(true);
-      }
-      else {
-        setIncludeComments(false);
-      }
-
-      setDebouncedSubreddit(subredditGo);
-      setSubreddit(subredditGo);
-      setSentimentKeywords(sentimentGo);
-      setStartDate(startDateGo);
-      setEndDate(endDateGo);
-      setOption(optionGo);
-
-      //issue where go to needs to be clicked twice
-      setTimeout(() => {
-        fetchArrowData(0, 10, 1, searchValue, setError).then(() => {
-          tableInitializedRef.current = true;
-          if ($.fn.DataTable.isDataTable("#click-table")) {
-            $("#click-table").DataTable().ajax.reload();
-          }
-          if (subredditGo) {
-            getSubredditIcon(subredditGo);
-          }
-        });
-      }, 50);
-
-    });
-
-    $("#clear-all-btn").on("click", ClearAllSearch);
-
-  }, [searchData]);
-
   const handleKeywordKeyDown = (e) => {
     if (e.key === 'Enter') {
       e.preventDefault();
@@ -548,9 +216,11 @@ function Data() {
               }
 
               if (apiData.data && apiData.data.length > 0) {
+                setDataLoadMessage(true);
                 getSubredditIcon(apiData.data[0].subreddit);
               }
               else {
+                setDataLoadMessage(false);
                 setSubredditIcon('../public/reddit-1.svg');
               }
             })
@@ -736,32 +406,51 @@ function Data() {
     setOption(selectedOptions.join(", "));
   }, [includeSubmissions, includeComments]);
 
+  //Issues with moving this to SearchHistory.jsx
+  const AddSearch = async () => {
+    try {
+      const response = await fetch('/api/add_search', {
+        method: 'POST',
+        headers: { 'Content-Type': 'application/json' },
+        body: JSON.stringify({
+          subreddit: `r/${subreddit}`,
+          sentimentKeywords: sentimentKeywords,
+          startDate: startDate,
+          endDate: endDate,
+          option: option,
+          email: email,
+        }),
+      });
+
+      if (!response.ok) {
+        setError('Failed to submit and add search.');
+      }
+      else {
+        setError(null);
+      }
+
+      const response2 = await fetch(`/api/get_search/${encodeURIComponent(email)}`);
+      const data2 = await response2.json();
+      setSearchData(data2.search_history || []);
+    } catch (error) {
+      setSearchError(error.message);
+    }
+  };
+
   // When Submit is clicked, mark table as initialized and reload the DataTable.
   const handleSubmit = async (e) => {
     e.preventDefault();
     setClusteringResults(null);
     // Build a comma-separated string, e.g., "reddit_submissions,reddit_comments"
-    const options = [];
-    if (includeSubmissions) options.push("reddit_submissions");
-    if (includeComments) options.push("reddit_comments");
-    const combinedOption = options.join(",");
-    setOption(combinedOption);
 
     setError(null);
     await AddSearch();  // (Make sure AddSearch also passes along the chosen option if needed.)
-    tableInitializedRef.current = true;
-    if ($.fn.DataTable.isDataTable("#click-table")) {
-      $("#click-table").DataTable().ajax.reload();
-    }
-    if (subreddit) {
-      getSubredditIcon(subreddit);
-    }
+
   };
 
   useEffect(() => {
     // Automatically run the search on page load. Adjust to only fetch data and not add to search history.
     const searchValue = `${subreddit} ${sentimentKeywords} ${startDate.toISOString()} ${endDate.toISOString()}`;
-
     fetchArrowData(0, 10, 1, searchValue, setError).then(() => {
       tableInitializedRef.current = true;
       if ($.fn.DataTable.isDataTable("#click-table")) {
@@ -841,17 +530,21 @@ function Data() {
     }
   };
 
-
-
   useEffect(() => {
     if (clusteringResults) {
       handleNotify();
     }
   }, [clusteringResults]);
 
-  const getSubredditIcon = async (subReddit) => {
+  const getSubredditIcon = async (subreddit) => {
     try {
-      const response = await fetch(`https://www.reddit.com/r/${subReddit}/about.json`);
+      //console.log('Test Icon')
+      if (subreddit === prevSubredditRef.current) {
+        //console.log('Same subreddit, skipping API call');
+        return;
+      }
+      //console.log('Called')
+      const response = await fetch(`https://www.reddit.com/r/${subreddit}/about.json`);
       const data = await response.json();
       let subredditIcon = data.data.community_icon;
 
@@ -860,13 +553,12 @@ function Data() {
         subredditIcon = subredditIcon.split('?')[0];
       }
 
-      //console.log('Cleaned Subreddit Icon:', subredditIcon);
-
       if (!subredditIcon) {
         setSubredditIcon('../public/reddit-1.svg');
       } else {
         setSubredditIcon(subredditIcon);
       }
+      prevSubredditRef.current = subreddit;
     } catch (error) {
       setSubredditIcon('../public/reddit-1.svg');
     }
@@ -1021,6 +713,7 @@ function Data() {
                 getSubredditIcon={getSubredditIcon}
                 isTyping={isTyping}
                 setIsTyping={setIsTyping}
+                dataLoadMessage={dataLoadMessage}
               />
             </div>
 
@@ -1253,27 +946,27 @@ function Data() {
       </div>
 
       {/* Search History DataTable */}
-      <div className="mt-5">
-        <h2><i className="fas fa-search me-1"></i> Search History</h2>
-        {searchError && <p className="text-danger mt-3">Error: {searchError}</p>}
-        <div id="clear-all-container" style={{ position: 'relative' }}>
-          <button id="clear-all-btn" className="btn btn-danger" style={{ position: 'absolute', right: '50px' }}>Clear All Searches</button>
-        </div>
-        <div>
-          <table id="search-history-table" className="display">
-            <thead>
-              <tr>
-                <th></th>
-                <th></th>
-                <th></th>
-                <th></th>
-                <th></th>
-              </tr>
-            </thead>
-            <tbody></tbody>
-          </table>
-        </div>
-      </div>
+      <SearchHistory
+        email={email}
+        setIncludeSubmissions={setIncludeSubmissions}
+        setIncludeComments={setIncludeComments}
+        setDebouncedSubreddit={setDebouncedSubreddit}
+        setSubreddit={setSubreddit}
+        setSentimentKeywords={setSentimentKeywords}
+        setStartDate={setStartDate}
+        setEndDate={setEndDate}
+        setOption={setOption}
+        fetchArrowData={fetchArrowData}
+        setError={setError}
+        tableInitializedRef={tableInitializedRef}
+        getSubredditIcon={getSubredditIcon}
+        searchData={searchData}
+        setSearchData={setSearchData}
+        searchError={searchError}
+        setSearchError={setSearchError}
+        historyInitializedRef={historyInitializedRef}
+        setIsDeleting={setIsDeleting}
+      />
     </div>
   );
 }
